Adapter new type promotion rule for Paddle 2.6 (#8421)

+++ model_zoo/gpt-3/ppfleetx/models/language_model/gpt/dygraph/single_model.py
@@ -809,7 +809,7 @@ class GPTForGeneration(nn.Layer):
     def update_scores_for_generation(self, scores, next_scores, length, unfinished_flag):
         # update scores
 
-        unfinished_scores = (scores * length + next_scores) / (length + 1)
+        unfinished_scores = (scores * length.astype(scores.dtype) + next_scores) / (length.astype(scores.dtype) + 1)
         scores = paddle.where(unfinished_flag, unfinished_scores, scores)
         return scores
